requirements.txt
openai==1.2.0
requests==2.31.0
pandas==2.0.3
click==8.1.3
google-cloud-aiplatform==1.49.0
mistral==18.0.1
mistral-lib==3.0.0
mistralai==0.1.8
vertexai==1.49.0
pydantic==2.7.1
pydantic-settings==2.0.2
pydantic_core==2.18.2
numpy==1.24.0
qwen_agent==0.0.10
python-dotenv==1.0.0
